+++ application/habiTv/src/com/dabi/habitv/HabitvLauncher.java
@@ -7,21 +7,26 @@ import org.apache.log4j.Logger;
 
 import com.dabi.habitv.console.ConsoleLauncher;
 import com.dabi.habitv.tray.HabiTvViewRunner;
-import com.dabi.habitv.utils.LogUtils;
+import com.dabi.habitv.utils.HabitvLogger;
 
 public class HabitvLauncher {
 
-	private static final Logger LOG = Logger.getLogger(HabitvLauncher.class);
+	private static final Logger LOG = HabitvLogger.getLogger(HabitvLauncher.class);
 
 	public static void main(final String[] args) throws Exception {
-		LogUtils.updateLog4jConfiguration();
+		// Initialize unified logging system early
+		HabitvLogger.initialize();
+		
 		Utils4J.addToClasspath("file:///" + System.getProperty("java.home")
 				+ File.separator + "lib" + File.separator + "jfxrt.jar");
-		LOG.info(System.getProperty("java.home"));
+		LOG.info("Java home: " + System.getProperty("java.home"));
+		LOG.info("habiTv application starting...");
 
 		if (args == null || args.length == 0) {
+			LOG.info("Starting GUI mode");
 			HabiTvViewRunner.main(args);
 		} else {
+			LOG.info("Starting console mode with " + args.length + " arguments");
 			ConsoleLauncher.main(args);
 		}
 	}